+++ R/resample_est.R
@@ -35,7 +35,7 @@
 #' # resample
 #' Y <- resample_est(X)
 #' }
-#' 
+#' @family extended selection table manipulation
 #' @seealso \code{\link{mp32wav}}, \code{\link{fix_wavs}}
 #' @references {
 #' Araya-Salas, M., & Smith-Vidaurre, G. (2017). warbleR: An R package to streamline analysis of animal acoustic signals. Methods in Ecology and Evolution, 8(2), 184-191.